+++ app/src/main/java/com/avinashdavid/readitforreddit/PostUtils/CommentRecord.java
@@ -29,7 +29,7 @@ public class CommentRecord extends SugarRecord {
 
     public String parent;
 
-    public float createdTime;
+    public long createdTime;
 
     public int depth;
 
@@ -42,7 +42,7 @@ public class CommentRecord extends SugarRecord {
     public CommentRecord() {
     }
 
-    public CommentRecord(long timestamp, String commentId, String linkId, boolean scoreHidden, int score, String author, String bodyHtml, String parent, float createdTime, int depth, boolean hasReplies, String authorFlairText, boolean isGilded, boolean isEdited) {
+    public CommentRecord(long timestamp, String commentId, String linkId, boolean scoreHidden, int score, String author, String bodyHtml, String parent, long createdTime, int depth, boolean hasReplies, String authorFlairText, boolean isGilded, boolean isEdited) {
         this.timestamp = timestamp;
         this.commentId = commentId;
         this.linkId = linkId;
